+++ CHANGELOG.md
@@ -4,9 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `color` prop to `HorizontalStroke`.
-- Fix: Remove `classNames` bug from `HorizontalStroke`.
-- Fix: Remove warning in `Pagination`.
+## [2.70.0] - 2020-05-15
+
+Feature:
+- Add `color` prop to `HorizontalStroke`.
+
+Fixes:
+- Remove `classNames` bug from `HorizontalStroke`.
+- Remove warning in `Pagination`.
 
 ## [2.69.0] - 2020-05-12